pyproject.toml
[tool.poetry]
name = "parcelle-watch"
version = "0.1.0"
description = "Surveillance satellite open source des parcelles agricoles"
authors = ["Votre Nom <votre.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Images satellite & géospatial
sentinelhub = "^3.10"
rasterio = "^1.3"
shapely = "^2.0"
geopandas = "^0.14"
pyproj = "^3.6"

# Data & calcul
numpy = "^1.26"
pandas = "^2.1"
xarray = "^2024.1"

# Météo
openmeteo-requests = "^1.2"
requests-cache = "^1.1"
retry-requests = "^2.0"

# Machine Learning
scikit-learn = "^1.4"
xgboost = "^2.0"

# Visualisation
folium = "^0.15"
plotly = "^5.18"
matplotlib = "^3.8"

# Dashboard
streamlit = "^1.31"
streamlit-folium = "^0.18"

# Rapport PDF
reportlab = "^4.1"
Pillow = "^10.2"

# Utilitaires
python-dotenv = "^1.0"
loguru = "^0.7"
tqdm = "^4.66"
click = "^8.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0"
ipykernel = "^6.29"
nbformat = "^5.9"
pytest = "^8.0"
pytest-cov = "^4.1"
black = "^24.0"
ruff = "^0.3"
mypy = "^1.8"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
